split files and rename
